make it RESTful + add checks for that id in url field equals to id sent in amandustoken
