+++ tokenservice/src/index.ts
@@ -1,6 +1,5 @@
 import express from 'express'
 import tokenRouter from './routes/tokenRouter'
-import userRouter from './routes/userRouter'
 import cors from 'cors'
 
 const app = express()
@@ -9,7 +8,6 @@ app.use(express.json())
 app.use(cors())
 
 app.use('/api/tokens', tokenRouter)
-app.use('/api/users', userRouter)
 
 const PORT = 3002
 app.listen(PORT, () => {
